pyproject.toml
[project]
name = "airflow-health-pipeline"
version = "0.1.0"
description = "Personal health data pipeline using MyFitnessPal and Garmin data"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==2.9.3",
    "apache-airflow-providers-postgres>=5.6.0",
    "pandas>=2.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []